added window overlap config setting - removed unsused changes
