+++ config.py
@@ -326,6 +326,7 @@ _C.INFERENCE.EVALUATION_METHOD = 'FFT'
 _C.INFERENCE.EVALUATION_WINDOW = CN()
 _C.INFERENCE.EVALUATION_WINDOW.USE_SMALLER_WINDOW = False
 _C.INFERENCE.EVALUATION_WINDOW.WINDOW_SIZE = 10
+_C.INFERENCE.EVALUATION_WINDOW.WINDOW_OVERLAP = 0
 _C.INFERENCE.MODEL_PATH = ''
 
 # -----------------------------------------------------------------------------
