+++ configs/infer_configs/MR-NIRP_UNSUPERVISED.yaml
@@ -1,7 +1,7 @@
 BASE: ['']
 TOOLBOX_MODE: "unsupervised_method"      # "train_and_test"  or "only_test" or "unsupervised_method"
 UNSUPERVISED:
-  METHOD: ["ICA", "POS", "CHROM", "GREEN", "LGI", "PBV"]
+  METHOD: ["ICA"]
   METRICS: ['MAE', 'RMSE', 'MAPE', 'Pearson', 'SNR']
   OUTPUT_SAVE_DIR: "./data"
   DATA:
@@ -39,3 +39,4 @@ INFERENCE:
   EVALUATION_WINDOW:
     USE_SMALLER_WINDOW: True        # Change this if you'd like an evaluation window smaller than the test video length
     WINDOW_SIZE: 10        # In seconds
+    WINDOW_OVERLAP: 100    # In frames
