+++ evaluation/post_process.py
@@ -120,53 +120,3 @@ def calculate_metric_per_video(predictions, labels, fs=30, diff_flag=True, use_b
         raise ValueError('Please use FFT or Peak to calculate your HR.')
     SNR = _calculate_SNR(predictions, hr_label, fs=fs)
     return hr_label, hr_pred, SNR
-
-
-# Resp Metrics
-def _calculate_fft_rr(resp_signal, fs=30, low_pass=0.13, high_pass=0.5):
-    """Calculate heart rate based on PPG using Fast Fourier transform (FFT)."""
-    resp_signal = np.expand_dims(resp_signal, 0)
-    N = _next_power_of_2(resp_signal.shape[1])
-    f_resp, pxx_resp = scipy.signal.periodogram(resp_signal, fs=fs, nfft=N, detrend=False)
-    fmask_resp = np.argwhere((f_resp >= low_pass) & (f_resp <= high_pass))
-    mask_resp = np.take(f_resp, fmask_resp)
-    mask_pxx = np.take(pxx_resp, fmask_resp)
-    fft_rr = np.take(mask_resp, np.argmax(mask_pxx, 0))[0] * 60
-    return fft_rr
-
-
-def _calculate_peak_rr(resp_signal, fs):
-    """Calculate heart rate based on PPG using peak detection."""
-    resp_peaks, _ = scipy.signal.find_peaks(resp_signal)
-    rr_peak = 60 / (np.mean(np.diff(resp_peaks)) / fs)
-    return rr_peak
-
-
-def calculate_resp_metrics_per_video(predictions, labels, fs=30, diff_flag=False, use_bandpass=True, rr_method='FFT'):
-    """Calculate video-level RR"""
-    if diff_flag:  # if the predictions and labels are 1st derivative of Resp signal.
-        predictions = _detrend(np.cumsum(predictions), 100)
-        labels = _detrend(np.cumsum(labels), 100)
-    else:
-        predictions = _detrend(predictions, 100)
-        labels = _detrend(labels, 100)
-        
-    
-    if use_bandpass:
-        low_pass = 0.13
-        high_pass = 0.5
-        # bandpass filter between [0.13, 0.5] Hz
-        # equals [8, 30] breaths per min
-        [b, a] = butter(1, [low_pass / fs * 2, high_pass / fs * 2], btype='bandpass')
-        predictions = scipy.signal.filtfilt(b, a, np.double(predictions))
-        labels = scipy.signal.filtfilt(b, a, np.double(labels))
-    if rr_method == 'FFT':
-        rr_pred = _calculate_fft_rr(predictions, fs=fs)
-        rr_label = _calculate_fft_rr(labels, fs=fs)
-    elif rr_method == 'Peak':
-        rr_pred = _calculate_peak_rr(predictions, fs=fs)
-        rr_label = _calculate_peak_rr(labels, fs=fs)
-    else:
-        raise ValueError('Please use FFT or Peak to calculate your RR.')
-    SNR = _calculate_SNR(predictions, rr_label, fs=fs, low_pass=0.13, high_pass=0.5)
-    return rr_label, rr_pred, SNR
